refactor: :recycle: refactor form to its own component
made add habit form its own component, called in add habit page
